Add and style flash message for successfully logging in, and fix bug
The bug was introduced by the previous commit:
Whenever a User failed to log in, and then logged in successfully,
they would be correctly redirected to their profile,
but the previous failed-login error message would be displayed.

This commit fixes that bug by adding a success message, but I should watch for similar bugs.
I may need to re-write the partial, or else call it only from the view files that display flash messages.