Start moving the responsibility for creating a new MyVerse and VerseReference to the MyVerse model
This is the best way to structure my project.
My goal is to allow users to CRUD my_verses, but only create and view verse_references.
I want users to be able to create (and ideally edit) a new my_verse and verse_reference with the same form.
I also want users to be able to choose between creating a new VerseReference or selecting one from a list.
I had thought that this responsibility belonged to the VerseReference, but it is better suited to the MyVerse.
These next few commits will transition that responsibility from one model to the other.