Replace nil values in seeds file with blank strings
As I was trying to create a new MyVerse, I discovered a bug: Duplicate VerseReferences were being created.
The reason for this is that when a VerseReference was created through a 'New MyVerse' form,
if the User didn't specify a verse_end, it would show up as "".
But VerseReferences made through the seed file showed up as nil.
Since nil is not "", similiar VerseReferences could be created. Thus, I altered the seed file to use "" instead of nil values.
By the same token, I figured that something similar would probably happen with MyVerses and Users,
so I changed that in the seed file as well.