Simplify the process of preventing a User from making a MyVerse for another User
Instead of providing a hidden field with a user_id,
checking whether that user_id is the current_user's id,
and re-rendering the "New MyVerse" page with an error if it isn't,
I simply used the MyVersesController to set the MyVerse's user_id to the current_user's id.
This uses less code, especially in the controller, and arguably makes the app safer.