Fix VerseReference default_scope bug and add notes and stretch goal
As noted in the VerseReference model, the original default_scope didn't completely work
because the :chapter, :verse_start, and :verse_end data types were strings, not integers;
they are not sorted in the same way. This is fixed for now,
but I should later refactor the verse_references table itself by changing the datatypes.
That should make my program faster by using less SQL in the VerseReference model.