Allow .spine element to scroll with the page
This way, instead of having a spine that cuts off at the bottom if the page is too tall in height,
it scrolls with the page instead, providing a more consistent appearance.

However, this removes the .spine element from the flex layout,
thus messing up the adjacent .page element by not centering its content.
I will attempt to fix that in the next commit.