Fix .page centering issue, and remove unnecessary CSS declaration
In order to fix the problem of the change to .spine messing up the centering of .page's content
(see previous commit), I added an empty <div> element in the same place as .spine and styled it with the same width.

Also, since the .spine element has been removed from the flexbox layout, the 'flex: 1;' CSS declaration is not needed.