Adjust spacing on MyVerses index page, and only display edit/delete buttons on user profile and MyVerse page
This partial (so far) is used by the MyVerses index and the user profile pages.
The change to the spacing is more noticeable on the index page. Now the links to the MyVerse and user profile page are on one line.
I also decided to remove the MyVerse edit/delete buttons on the index page, for a more consistent and neater layout;
they are still on the MyVerse page and user profile page if the current user created the MyVerse.